+++ includes/widgets/mybooking-plugin-rent-selector-wizard-widget.php
@@ -37,6 +37,13 @@ class MyBookingRentEngineSelectorWizardWidget extends WP_Widget {
             $data['family_id'] = '';
         }
 
+        if ( array_key_exists('category_code', $instance) ) {
+            $data['category_code'] = $instance['category_code'];
+        }
+        else {
+            $data['category_code'] = '';
+        }
+
         if ( array_key_exists('rental_location_code', $instance) ) {
             $data['rental_location_code'] = $instance['rental_location_code'];
         }
@@ -76,7 +83,19 @@ class MyBookingRentEngineSelectorWizardWidget extends WP_Widget {
                     name="<?php echo esc_attr( $this->get_field_name( 'family_id' ) ); ?>" 
                     type="text" 
                     value="<?php echo esc_attr( $family_id ); ?>">
-            </p>    
+            </p>  
+            <p>
+                <label for="<?php echo esc_attr( $this->get_field_id( 'category_code' ) ); ?>">
+                <?php esc_attr_e( 'category_code:', 'text_domain' ); ?>
+                </label> 
+                
+                <input 
+                    class="widefat" 
+                    id="<?php echo esc_attr( $this->get_field_id( 'category_code' ) ); ?>" 
+                    name="<?php echo esc_attr( $this->get_field_name( 'category_code' ) ); ?>" 
+                    type="text" 
+                    value="<?php echo esc_attr( $category_code ); ?>">
+            </p>               
             <p>
                 <label for="<?php echo esc_attr( $this->get_field_id( 'rental_location_code' ) ); ?>">
                 <?php esc_attr_e( 'rental_location_code:', 'text_domain' ); ?>
@@ -97,6 +116,7 @@ class MyBookingRentEngineSelectorWizardWidget extends WP_Widget {
         $instance = array();
         $instance['sales_channel_code'] = ( ! empty( $new_instance['sales_channel_code'] ) ) ? strip_tags( $new_instance['sales_channel_code'] ) : '';
         $instance['family_id'] = ( ! empty( $new_instance['family_id'] ) ) ? strip_tags( $new_instance['family_id'] ) : '';
+        $instance['category_code'] = ( ! empty( $new_instance['category_code'] ) ) ? strip_tags( $new_instance['category_code'] ) : '';
         $instance['rental_location_code'] = ( ! empty( $new_instance['rental_location_code'] ) ) ? strip_tags( $new_instance['rental_location_code'] ) : '';
         return $instance;        
     }